[Refactor][Zeta] Promote savepoint to a peer package of checkpoint
The savepoint feature owns its storage capability, bundle commit protocol
and its own versioned wire format, so it is no longer a sub-detail of the
checkpoint runtime. Move it to engine.server.savepoint (mirroring Flink's
runtime.checkpoint / runtime.savepoint split):

- engine.server.checkpoint            # checkpoint runtime (unchanged)
- engine.server.savepoint             # savepoint feature layer
  - SavepointSaveSession
  - serialization                     # engine-wire v1 DTOs, codec, legacy
                                       reader, versioned readers/registry

The checkpoint runtime still interacts with savepoints only through the
storage SPI (SavepointWriter) and the reader registry.
