+++ seatunnel-engine/seatunnel-engine-server/src/main/java/org/apache/seatunnel/engine/server/savepoint/serialization/WireSubtaskState.java
@@ -15,32 +15,26 @@
  * limitations under the License.
  */
 
-package org.apache.seatunnel.engine.server.checkpoint.savepoint.serialization;
+package org.apache.seatunnel.engine.server.savepoint.serialization;
 
 import io.protostuff.Tag;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-/** Wire-format DTO of one subtask statistics entry ({@code engine-wire-v1}). */
+import java.util.List;
+
+/** Wire-format DTO of one subtask state ({@code engine-wire-v1}). */
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class WireSubtaskStatistics {
+public class WireSubtaskState {
 
+    /** Coordinator subtasks use {@code index == -1}. */
     @Tag(1)
-    private int subtaskIndex;
+    private int index;
 
+    /** Connector state bytes; engine treats them as opaque. */
     @Tag(2)
-    private long ackTimestamp;
-
-    @Tag(3)
-    private long stateSize;
-
-    /**
-     * Stable {@link org.apache.seatunnel.engine.server.checkpoint.SubtaskStatus#name()} value; must
-     * not rely on enum ordinal.
-     */
-    @Tag(4)
-    private String subtaskStatusName;
+    private List<byte[]> state;
 }
