+++ seatunnel-engine/seatunnel-engine-server/src/test/java/org/apache/seatunnel/engine/server/savepoint/SavepointRestoreSelectionTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.seatunnel.engine.server.checkpoint.savepoint;
+package org.apache.seatunnel.engine.server.savepoint;
 
 import org.apache.seatunnel.common.utils.FileUtils;
 import org.apache.seatunnel.engine.checkpoint.storage.api.CheckpointStorage;
